+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Drop uneeded `ollama useTools` and `ollama think` configs.
 - Refactor configs for config providers unification.
   - `<provider>ApiKey` and `<providerApiUrl>` now live in `:providers "<provider>" :key`.
+  - Move `defaultModel` config from customProvider to root. 
 
 ## 0.30.0